refactor: 에러 핸들링 개선
