+++ src/js/components/Login/Login.js
@@ -4,6 +4,7 @@ import {
   API,
   showSnackbar,
   setLocalStorageItem,
+  getLocalStorageItem,
 } from '../../utils/index.js';
 import { loginTemplate } from './template.js';
 import { SNACKBAR_MESSAGE, LOCAL_STORAGE_KEY } from '../../constants/index.js';
@@ -32,23 +33,26 @@ export default class Login extends Component {
     const email = e.target.elements['login-email'].value;
     const password = e.target.elements['login-password'].value;
 
-    const response = await API.login({ email, password });
-    const responseJSON = await response.json();
+    try {
+      const response = await API.login({ email, password });
+      const responseJSON = await response.json();
 
-    if (!responseJSON.accessToken) {
+      if (!responseJSON.accessToken) {
+        return;
+      }
+
+      showSnackbar(SNACKBAR_MESSAGE.LOGIN_SUCCESS);
+      setLocalStorageItem({
+        key: LOCAL_STORAGE_KEY.TOKEN,
+        item: responseJSON.accessToken,
+      });
+
+      this.changeTemplate('/');
+      history.pushState({ pathName: '/' }, null, '/');
+      Navigation.changeSelectedButtonColor();
+    } catch {
       showSnackbar(SNACKBAR_MESSAGE.LOGIN_FAILURE);
-      return;
     }
-
-    showSnackbar(SNACKBAR_MESSAGE.LOGIN_SUCCESS);
-    setLocalStorageItem({
-      key: LOCAL_STORAGE_KEY.TOKEN,
-      item: responseJSON.accessToken,
-    });
-
-    this.changeTemplate('/');
-    history.pushState({ pathName: '/' }, null, '/');
-    Navigation.changeSelectedButtonColor();
   }
 
   handleSingupButton(e) {
