+++ src/js/constants/index.js
@@ -1,3 +1,3 @@
-export { SNACKBAR_MESSAGE } from './snackbarMessage.js';
 export { LOCAL_STORAGE_KEY } from './standard.js';
 export { LOGIN_REQUIRED_TEMPLATE } from './template.js';
+export { MESSAGE, SNACKBAR_MESSAGE } from './message.js';
