+++ src/js/index.js
@@ -1,4 +1,4 @@
-import '../css/index.css';
+// import '../css/index.css';
 import App from './components/App.js';
 
 window.addEventListener('DOMContentLoaded', () => {
